Share ShortcutChip/ShortcutKbd and open-*-page utils across popup and readability
Moves components/utils that both features independently depended on
(popup reaching into readability's folder) into feature-agnostic
homes under src/components and src/utils, reachable via a new
@stylebot/components alias alongside the existing @stylebot/utils
one. Also moves the popup's switch color CSS variables out of App.vue's
global :root into ToggleSwitch.vue itself, since it was their only
consumer.
