+++ src/components/ShortcutChip.vue
