+++ src/components/ShortcutKbd.vue
