+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as ShortcutChip } from './ShortcutChip.vue';
+export { default as ShortcutKbd } from './ShortcutKbd.vue';
