+++ src/components/utils/__tests__/format-shortcut.test.ts
