+++ src/popup/components/Style.vue
@@ -29,7 +29,7 @@ import Vue from 'vue';
 import { EnableStyle, DisableStyle } from '@stylebot/types';
 import PopupRow from './PopupRow.vue';
 import ToggleSwitch from './ToggleSwitch.vue';
-import ShortcutChip from '../../readability/components/dock/ShortcutChip.vue';
+import { ShortcutChip } from '@stylebot/components';
 
 export default Vue.extend({
   name: 'Style',
