--- src/readability/utils/open-donate-page.ts
@@ -1,6 +1,0 @@
-/**
- * Opens the Ko-fi donation page in a new tab.
- */
-export const openDonatePage = (): void => {
-  chrome.runtime.sendMessage({ name: 'OpenDonatePage' });
-};
