--- src/readability/utils/open-options-page.ts
@@ -1,6 +1,0 @@
-/**
- * Opens the Stylebot options page in a new tab.
- */
-export const openOptionsPage = (): void => {
-  chrome.runtime.sendMessage({ name: 'OpenOptionsPage' });
-};
