--- src/readability/utils/open-report-issue-page.ts
@@ -1,6 +1,0 @@
-/**
- * Opens the GitHub issue tracker in a new tab.
- */
-export const openReportIssuePage = (): void => {
-  chrome.runtime.sendMessage({ name: 'OpenReportIssuePage' });
-};
